refactor(frontend/lobby): moved lobby check to play_panel.js
