+++ frontend/src/js/play_panel.js
@@ -4,6 +4,8 @@ import { loadPage } from "./router.js";
 import { loadContentToTarget } from "./ui_utils/ui_utils.js";
 import { getRequest } from "./network_utils/api_requests.js";
 import { PanelSwitcher } from "./ui_utils/panelswitcher_utils.js";
+import { initClassicLobby, updateClassicLobby, initTournamentLobby, updateTournamentLobby, getLobbyType, checkInLobby } from "./lobby.js";
+import { checkInTournament } from "./tournament.js";
 
 // for moving stuff
 export var startingMenuSwitcher = null
@@ -63,4 +65,31 @@ export async function startLocalGame() {
   }
   await loadPage('game')
   joinMatch(matchID, defaultMatchOnClose)
+}
+
+export async function tryReturnToLobby() {
+  // resume to lobby
+  if (!checkInLobby()) {
+    return
+  }
+
+  divSwitcher.disableDivInput('play-select-container')
+  if (!checkInTournament()) {
+    const inTournament = (getLobbyType() == 'tournament')
+    if (inTournament) {
+      await loadContentToTarget('menu/lobby_tournament_content.html', 'play-lobby-container')
+      initTournamentLobby()
+    } else {
+      await loadContentToTarget('menu/lobby_classic_content.html', 'play-lobby-container')
+      initClassicLobby()
+    }
+
+    divSwitcher.setCurrentDiv('play-select-container', 'play-lobby-container', true)
+    if (inTournament) {
+      updateTournamentLobby()
+    } else {
+      updateClassicLobby()
+    }
+  } else {
+  }
 }
